Rearrange the methods of the backend class before I lose my mind.
