+++ lsif/src/server/backend/backend.ts
@@ -11,90 +11,7 @@ import { mustGet } from '../../shared/maps'
 import { DumpManager } from '../../shared/store/dumps'
 import { DependencyManager } from '../../shared/store/dependencies'
 import { isDefined } from '../../shared/util'
-
-/**
- * Context describing the current request for paginated results.
- */
-export interface ReferencePaginationContext {
-    /**
-     * The maximum number of remote dumps to search.
-     */
-    limit: number
-
-    /**
-     * Context describing the previous page of results.
-     */
-    cursor?: ReferencePaginationCursor
-}
-
-/**
- * Reference pagination happens in two distinct phases:
- *
- *   (1) open a slice of dumps for the same repositories, and
- *   (2) open a slice of dumps for other repositories.
- */
-export type ReferencePaginationPhase = 'same-repo' | 'remote-repo'
-
-/**
- * Context describing the previous page of results.
- */
-export interface ReferencePaginationCursor {
-    /**
-     * The identifier of the dump that contains the target range.
-     */
-    dumpId: number
-
-    /**
-     * The scheme of the moniker that has remote results.
-     */
-    scheme: string
-
-    /**
-     * The identifier of the moniker that has remote results.
-     */
-    identifier: string
-
-    /**
-     * The name of the package that has remote results.
-     */
-    name: string
-
-    /**
-     * The version of the package that has remote results.
-     */
-    version: string | null
-
-    /**
-     * The phase of the pagination.
-     */
-    phase: ReferencePaginationPhase
-
-    /**
-     * The number of remote dumps to skip.
-     */
-    offset: number
-}
-
-/**
- * Converts a file in the repository to the corresponding file in the
- * database.
- *
- * @param root The root of all files in the dump.
- * @param path The path within the dump.
- */
-const pathToDatabase = (root: string, path: string): string => (path.startsWith(root) ? path.slice(root.length) : path)
-
-/**
- * Converts a location in a dump to the corresponding location in the repository.
- *
- * @param root The root of all files in the dump.
- * @param location The original location.
- */
-const locationFromDatabase = (root: string, { dump, path, range }: InternalLocation): InternalLocation => ({
-    dump,
-    path: `${root}${path}`,
-    range,
-})
+import { ReferencePaginationContext, ReferencePaginationCursor } from './cursor'
 
 /**
  * A wrapper around code intelligence operations.
@@ -136,6 +53,10 @@ export class Backend {
         return (await this.findClosestDatabases(repositoryId, commit, path, ctx)).map(({ dump }) => dump)
     }
 
+    //
+    // DEFINITIONS
+    //
+
     /**
      * Return the location for the symbol at the given position. Returns undefined if no dump can
      * be loaded to answer this query.
@@ -161,82 +82,6 @@ export class Backend {
         }
 
         return result.locations
-    }
-
-    /**
-     * Return a list of locations which reference the symbol at the given position. Returns
-     * undefined if no dump can be loaded to answer this query.
-     *
-     * @param repositoryId The repository identifier.
-     * @param commit The commit.
-     * @param path The path of the document to which the position belongs.
-     * @param position The current hover position.
-     * @param paginationContext Context describing the current request for paginated results.
-     * @param dumpId The identifier of the dump to load. If not supplied, the closest dump will be used.
-     * @param ctx The tracing context.
-     */
-    public async references(
-        repositoryId: number,
-        commit: string,
-        path: string,
-        position: lsp.Position,
-        paginationContext: ReferencePaginationContext = { limit: 10 },
-        dumpId?: number,
-        ctx: TracingContext = {}
-    ): Promise<{ locations: InternalLocation[]; newCursor?: ReferencePaginationCursor } | undefined> {
-        return this.internalReferences(repositoryId, commit, path, position, paginationContext, dumpId, ctx)
-    }
-
-    /**
-     * Return the hover content for the symbol at the given position. Returns undefined if no dump can
-     * be loaded to answer this query.
-     *
-     * @param repositoryId The repository identifier.
-     * @param commit The commit.
-     * @param path The path of the document to which the position belongs.
-     * @param position The current hover position.
-     * @param dumpId The identifier of the dump to load. If not supplied, the closest dump will be used.
-     * @param ctx The tracing context.
-     */
-    public async hover(
-        repositoryId: number,
-        commit: string,
-        path: string,
-        position: lsp.Position,
-        dumpId?: number,
-        ctx: TracingContext = {}
-    ): Promise<{ text: string; range: lsp.Range } | null | undefined> {
-        const closestDatabaseAndDump = await this.closestDatabase(repositoryId, commit, path, dumpId, ctx)
-        if (!closestDatabaseAndDump) {
-            if (ctx.logger) {
-                ctx.logger.warn('No database could be loaded', { repositoryId, commit, path })
-            }
-
-            return undefined
-        }
-        const { database, dump, ctx: newCtx } = closestDatabaseAndDump
-
-        // Try to find hover in the same dump
-        const hover = await database.hover(pathToDatabase(dump.root, path), position, newCtx)
-        if (hover !== null) {
-            return hover
-        }
-
-        // If we don't have a local hover, lookup the definitions of the
-        // range and read the hover data from the remote database. This
-        // can happen when the indexer only gives a moniker but does not
-        // give hover data for externally defined symbols.
-
-        const result = await this.internalDefinitions(repositoryId, commit, path, position, dumpId, ctx)
-        if (result === undefined || result.locations.length === 0) {
-            return null
-        }
-
-        return this.createDatabase(result.locations[0].dump).hover(
-            pathToDatabase(result.locations[0].dump.root, result.locations[0].path),
-            result.locations[0].range.start,
-            newCtx
-        )
     }
 
     private async internalDefinitions(
@@ -318,7 +163,23 @@ export class Backend {
         return { dump, locations: [] }
     }
 
-    private async internalReferences(
+    //
+    // REFERENCES
+    //
+
+    /**
+     * Return a list of locations which reference the symbol at the given position. Returns
+     * undefined if no dump can be loaded to answer this query.
+     *
+     * @param repositoryId The repository identifier.
+     * @param commit The commit.
+     * @param path The path of the document to which the position belongs.
+     * @param position The current hover position.
+     * @param paginationContext Context describing the current request for paginated results.
+     * @param dumpId The identifier of the dump to load. If not supplied, the closest dump will be used.
+     * @param ctx The tracing context.
+     */
+    public async references(
         repositoryId: number,
         commit: string,
         path: string,
@@ -326,11 +187,9 @@ export class Backend {
         paginationContext: ReferencePaginationContext = { limit: 10 },
         dumpId?: number,
         ctx: TracingContext = {}
-    ): Promise<
-        { dump: pgModels.LsifDump; locations: InternalLocation[]; newCursor?: ReferencePaginationCursor } | undefined
-    > {
+    ): Promise<{ locations: InternalLocation[]; newCursor?: ReferencePaginationCursor } | undefined> {
         if (paginationContext.cursor) {
-            return this.handleReferencesNextPage(
+            return this.handleReferencePaginationCursor(
                 repositoryId,
                 commit,
                 paginationContext.limit,
@@ -362,7 +221,7 @@ export class Backend {
 
         const { document, ranges } = await database.getRangeByPosition(pathInDb, position, ctx)
         if (!document || ranges.length === 0) {
-            return { dump, locations: [] }
+            return { locations: [] }
         }
 
         for (const range of ranges) {
@@ -430,7 +289,6 @@ export class Backend {
                 )
 
                 return {
-                    dump,
                     // TODO - determine source of duplication
                     locations: uniqWith(locations.concat(remoteLocations), isEqual),
                     newCursor,
@@ -439,118 +297,7 @@ export class Backend {
         }
 
         // TODO - determine source of duplication
-        return { dump, locations: uniqWith(locations, isEqual) }
-    }
-
-    /**
-     * Handle a references request given a non-empty pagination cursor.
-     *
-     * @param repositoryId The repository identifier.
-     * @param commit The commit.
-     * @param limit The pagination limit.
-     * @param cursor The pagination cursor.
-     * @param ctx The tracing context.
-     */
-    private async handleReferencesNextPage(
-        repositoryId: number,
-        commit: string,
-        limit: number,
-        cursor: ReferencePaginationCursor,
-        ctx: TracingContext = {}
-    ): Promise<
-        { dump: pgModels.LsifDump; locations: InternalLocation[]; newCursor?: ReferencePaginationCursor } | undefined
-    > {
-        const dump = await this.dumpManager.getDumpById(cursor.dumpId)
-        if (dump === undefined) {
-            return undefined
-        }
-
-        // Continue from previous page
-        const results = await this.handleReferencePaginationCursor(repositoryId, commit, limit, cursor, ctx)
-        if (results !== undefined) {
-            return { dump, ...results }
-        }
-
-        // No results remaining
-        return { dump, locations: [] }
-    }
-
-    /**
-     * Retrieve the package information from associated with the given moniker.
-     *
-     * @param document The document containing an instance of the moniker.
-     * @param moniker The target moniker.
-     * @param ctx The tracing context.
-     */
-    private lookupPackageInformation(
-        document: sqliteModels.DocumentData,
-        moniker: sqliteModels.MonikerData,
-        ctx: TracingContext = {}
-    ): sqliteModels.PackageInformationData | undefined {
-        if (!moniker.packageInformationId) {
-            return undefined
-        }
-
-        const packageInformation = document.packageInformation.get(moniker.packageInformationId)
-        if (!packageInformation) {
-            return undefined
-        }
-
-        logSpan(ctx, 'package_information', {
-            moniker,
-            packageInformation,
-        })
-
-        return packageInformation
-    }
-
-    /**
-     * Find the locations attached to the target moniker outside of the current database. If
-     * the moniker has attached package information, then Postgres is queried for the target
-     * package. That database is opened, and its definitions table is queried for the target
-     * moniker.
-     *
-     * @param document The document containing the definition.
-     * @param moniker The target moniker.
-     * @param model The target model.
-     * @param pagination A limit and offset to use for the query.
-     * @param ctx The tracing context.
-     */
-    private async lookupMoniker(
-        document: sqliteModels.DocumentData,
-        moniker: sqliteModels.MonikerData,
-        model: typeof sqliteModels.DefinitionModel | typeof sqliteModels.ReferenceModel,
-        pagination: { skip?: number; take?: number },
-        ctx: TracingContext = {}
-    ): Promise<InternalLocation[]> {
-        const packageInformation = this.lookupPackageInformation(document, moniker, ctx)
-        if (!packageInformation) {
-            return []
-        }
-
-        const packageEntity = await this.dependencyManager.getPackage(
-            moniker.scheme,
-            packageInformation.name,
-            packageInformation.version
-        )
-        if (!packageEntity) {
-            return []
-        }
-
-        logSpan(ctx, 'package_entity', {
-            moniker,
-            packageInformation,
-            packageRepositoryId: packageEntity.dump.repositoryId,
-            packageCommit: packageEntity.dump.commit,
-        })
-
-        const { locations: monikerResults } = await this.createDatabase(packageEntity.dump).monikerResults(
-            model,
-            moniker,
-            pagination,
-            ctx
-        )
-        return monikerResults.map(loc => locationFromDatabase(packageEntity.dump.root, loc))
+        return { locations: uniqWith(locations, isEqual) }
     }
 
     /**
@@ -630,8 +377,12 @@ export class Backend {
     }
 
     /**
-     * Perform a remote reference lookup on the dumps of the same repository. If there are any
-     * locations in the result set, this method returns the new cursor. This method return
+     *
+     * Find the references of the target moniker outside of the current dump but within a dump of
+     * the same repository. If the moniker has attached package information, then the dependency
+     * database is queried for the packages that require this particular moniker identifier. These
+     * dumps are opened, and their references tables are queried for the target moniker. If there
+     * are any locations in the result set, this method returns the new cursor. This method return
      * undefined if there are no remaining results for the same repository.
      *
      * @param repositoryId The repository identifier.
@@ -647,17 +398,17 @@ export class Backend {
         cursor: ReferencePaginationCursor,
         ctx: TracingContext = {}
     ): Promise<{ locations: InternalLocation[]; newCursor?: ReferencePaginationCursor }> {
-        const moniker = { scheme: cursor.scheme, identifier: cursor.identifier }
-        const packageInformation = { name: cursor.name, version: cursor.version }
-
-        const { locations, totalCount, newOffset } = await this.sameRepositoryRemoteReferences(
-            cursor.dumpId,
+        const { references, totalCount, newOffset } = await this.dependencyManager.getSameRepoRemoteReferences({
+            ...cursor,
             repositoryId,
             commit,
-            moniker,
-            packageInformation,
             limit,
-            cursor.offset,
+        })
+
+        const locations = await this.locationsFromRemoteReferences(
+            cursor.dumpId,
+            { scheme: cursor.scheme, identifier: cursor.identifier },
+            references.map(r => r.dump),
             ctx
         )
 
@@ -668,9 +419,12 @@ export class Backend {
     }
 
     /**
-     * Perform a remote reference lookup on dumps of remote repositories. If there are any
-     * locations in the result set, this method returns the new cursor. This method return
-     * undefined if there are no remaining results for the same repository.
+     * Find the references of the target moniker outside of the current repository. If the moniker
+     * has attached package information, then Postgres is queried for the packages that require
+     * this particular moniker identifier. These dumps are opened, and their references tables are
+     * queried for the target moniker. If there are any locations in the result set, this method
+     * returns the new cursor. This method return undefined if there are no remaining results for
+     * the same repository.
      *
      * @param repositoryId The repository identifier.
      * @param limit The maximum number of dumps to open.
@@ -683,16 +437,16 @@ export class Backend {
         cursor: ReferencePaginationCursor,
         ctx: TracingContext = {}
     ): Promise<{ locations: InternalLocation[]; newCursor?: ReferencePaginationCursor }> {
-        const moniker = { scheme: cursor.scheme, identifier: cursor.identifier }
-        const packageInformation = { name: cursor.name, version: cursor.version }
-
-        const { locations, totalCount, newOffset } = await this.remoteReferences(
-            cursor.dumpId,
+        const { references, totalCount, newOffset } = await this.dependencyManager.getReferences({
+            ...cursor,
             repositoryId,
-            moniker,
-            packageInformation,
             limit,
-            cursor.offset,
+        })
+
+        const locations = await this.locationsFromRemoteReferences(
+            cursor.dumpId,
+            { scheme: cursor.scheme, identifier: cursor.identifier },
+            references.map(r => r.dump),
             ctx
         )
 
@@ -700,85 +454,6 @@ export class Backend {
             locations,
             newCursor: newOffset < totalCount ? { ...cursor, phase: 'remote-repo', offset: newOffset } : undefined,
         }
-    }
-
-    /**
-     * Find the references of the target moniker outside of the current repository. If the moniker
-     * has attached package information, then Postgres is queried for the packages that require
-     * this particular moniker identifier. These dumps are opened, and their references tables are
-     * queried for the target moniker.
-     *
-     * @param dumpId The ID of the dump for which this database answers queries.
-     * @param repositoryId The repository identifier for which this database answers queries.
-     * @param moniker The target moniker.
-     * @param packageInformation The target package.
-     * @param limit The maximum number of remote dumps to search.
-     * @param offset The number of remote dumps to skip.
-     * @param ctx The tracing context.
-     */
-    private async remoteReferences(
-        dumpId: pgModels.DumpId,
-        repositoryId: number,
-        moniker: Pick<sqliteModels.MonikerData, 'scheme' | 'identifier'>,
-        packageInformation: Pick<sqliteModels.PackageInformationData, 'name' | 'version'>,
-        limit: number,
-        offset: number,
-        ctx: TracingContext = {}
-    ): Promise<{ locations: InternalLocation[]; totalCount: number; newOffset: number }> {
-        const { references, totalCount, newOffset } = await this.dependencyManager.getReferences({
-            repositoryId,
-            scheme: moniker.scheme,
-            identifier: moniker.identifier,
-            name: packageInformation.name,
-            version: packageInformation.version,
-            limit,
-            offset,
-        })
-
-        const dumps = references.map(r => r.dump)
-        const locations = await this.locationsFromRemoteReferences(dumpId, moniker, dumps, ctx)
-        return { locations, totalCount, newOffset }
-    }
-
-    /**
-     * Find the references of the target moniker outside of the current dump but within a dump of
-     * the same repository. If the moniker has attached package information, then the dependency
-     * database is queried for the packages that require this particular moniker identifier. These
-     * dumps are opened, and their references tables are queried for the target moniker.
-     *
-     * @param dumpId The ID of the dump for which this database answers queries.
-     * @param repositoryId The repository identifier for which this database answers queries.
-     * @param commit The commit of the references query.
-     * @param moniker The target moniker.
-     * @param packageInformation The target package.
-     * @param limit The maximum number of remote dumps to search.
-     * @param offset The number of remote dumps to skip.
-     * @param ctx The tracing context.
-     */
-    private async sameRepositoryRemoteReferences(
-        dumpId: pgModels.DumpId,
-        repositoryId: number,
-        commit: string,
-        moniker: Pick<sqliteModels.MonikerData, 'scheme' | 'identifier'>,
-        packageInformation: Pick<sqliteModels.PackageInformationData, 'name' | 'version'>,
-        limit: number,
-        offset: number,
-        ctx: TracingContext = {}
-    ): Promise<{ locations: InternalLocation[]; totalCount: number; newOffset: number }> {
-        const { references, totalCount, newOffset } = await this.dependencyManager.getSameRepoRemoteReferences({
-            repositoryId,
-            commit,
-            scheme: moniker.scheme,
-            identifier: moniker.identifier,
-            name: packageInformation.name,
-            version: packageInformation.version,
-            limit,
-            offset,
-        })
-
-        const dumps = references.map(r => r.dump)
-        const locations = await this.locationsFromRemoteReferences(dumpId, moniker, dumps, ctx)
-        return { locations, totalCount, newOffset }
     }
 
     /**
@@ -819,6 +494,148 @@ export class Backend {
 
         return locations
     }
+
+    //
+    // HOVER
+    //
+
+    /**
+     * Return the hover content for the symbol at the given position. Returns undefined if no dump can
+     * be loaded to answer this query.
+     *
+     * @param repositoryId The repository identifier.
+     * @param commit The commit.
+     * @param path The path of the document to which the position belongs.
+     * @param position The current hover position.
+     * @param dumpId The identifier of the dump to load. If not supplied, the closest dump will be used.
+     * @param ctx The tracing context.
+     */
+    public async hover(
+        repositoryId: number,
+        commit: string,
+        path: string,
+        position: lsp.Position,
+        dumpId?: number,
+        ctx: TracingContext = {}
+    ): Promise<{ text: string; range: lsp.Range } | null | undefined> {
+        const closestDatabaseAndDump = await this.closestDatabase(repositoryId, commit, path, dumpId, ctx)
+        if (!closestDatabaseAndDump) {
+            if (ctx.logger) {
+                ctx.logger.warn('No database could be loaded', { repositoryId, commit, path })
+            }
+
+            return undefined
+        }
+        const { database, dump, ctx: newCtx } = closestDatabaseAndDump
+
+        // Try to find hover in the same dump
+        const hover = await database.hover(pathToDatabase(dump.root, path), position, newCtx)
+        if (hover !== null) {
+            return hover
+        }
+
+        // If we don't have a local hover, lookup the definitions of the
+        // range and read the hover data from the remote database. This
+        // can happen when the indexer only gives a moniker but does not
+        // give hover data for externally defined symbols.
+
+        const result = await this.internalDefinitions(repositoryId, commit, path, position, dumpId, ctx)
+        if (result === undefined || result.locations.length === 0) {
+            return null
+        }
+
+        return this.createDatabase(result.locations[0].dump).hover(
+            pathToDatabase(result.locations[0].dump.root, result.locations[0].path),
+            result.locations[0].range.start,
+            newCtx
+        )
+    }
+
+    //
+    // MONIKER HELPERS
+    //
+
+    /**
+     * Find the locations attached to the target moniker outside of the current database. If
+     * the moniker has attached package information, then Postgres is queried for the target
+     * package. That database is opened, and its definitions table is queried for the target
+     * moniker.
+     *
+     * @param document The document containing the definition.
+     * @param moniker The target moniker.
+     * @param model The target model.
+     * @param pagination A limit and offset to use for the query.
+     * @param ctx The tracing context.
+     */
+    private async lookupMoniker(
+        document: sqliteModels.DocumentData,
+        moniker: sqliteModels.MonikerData,
+        model: typeof sqliteModels.DefinitionModel | typeof sqliteModels.ReferenceModel,
+        pagination: { skip?: number; take?: number },
+        ctx: TracingContext = {}
+    ): Promise<InternalLocation[]> {
+        const packageInformation = this.lookupPackageInformation(document, moniker, ctx)
+        if (!packageInformation) {
+            return []
+        }
+
+        const packageEntity = await this.dependencyManager.getPackage(
+            moniker.scheme,
+            packageInformation.name,
+            packageInformation.version
+        )
+        if (!packageEntity) {
+            return []
+        }
+
+        logSpan(ctx, 'package_entity', {
+            moniker,
+            packageInformation,
+            packageRepositoryId: packageEntity.dump.repositoryId,
+            packageCommit: packageEntity.dump.commit,
+        })
+
+        const { locations: monikerResults } = await this.createDatabase(packageEntity.dump).monikerResults(
+            model,
+            moniker,
+            pagination,
+            ctx
+        )
+        return monikerResults.map(loc => locationFromDatabase(packageEntity.dump.root, loc))
+    }
+
+    /**
+     * Retrieve the package information from associated with the given moniker.
+     *
+     * @param document The document containing an instance of the moniker.
+     * @param moniker The target moniker.
+     * @param ctx The tracing context.
+     */
+    private lookupPackageInformation(
+        document: sqliteModels.DocumentData,
+        moniker: sqliteModels.MonikerData,
+        ctx: TracingContext = {}
+    ): sqliteModels.PackageInformationData | undefined {
+        if (!moniker.packageInformationId) {
+            return undefined
+        }
+
+        const packageInformation = document.packageInformation.get(moniker.packageInformationId)
+        if (!packageInformation) {
+            return undefined
+        }
+
+        logSpan(ctx, 'package_information', {
+            moniker,
+            packageInformation,
+        })
+
+        return packageInformation
+    }
+
+    //
+    // DATABASE HELPERS
+    //
 
     /**
      * Create a database instance for the dump identifier. This identifier should have ben retrieved
@@ -916,5 +733,30 @@ export class Backend {
             dump,
             dbFilename(this.storageRoot, dump.id)
         )
+    }
+}
+
+/**
+ * Converts a file in the repository to the corresponding file in the
+ * database.
+ *
+ * @param root The root of all files in the dump.
+ * @param path The path within the dump.
+ */
+function pathToDatabase(root: string, path: string): string {
+    return path.startsWith(root) ? path.slice(root.length) : path
+}
+
+/**
+ * Converts a location in a dump to the corresponding location in the repository.
+ *
+ * @param root The root of all files in the dump.
+ * @param location The original location.
+ */
+function locationFromDatabase(root: string, { dump, path, range }: InternalLocation): InternalLocation {
+    return {
+        dump,
+        path: `${root}${path}`,
+        range,
     }
 }
